Cleanup of threadData.
Removed the stack field (because pthreads now provides the
stacks) and the dynamic scope field (because we've been
storing that in the continuation vector for a while now),
and replaced all use of "edenRoot" with "shelter" (formerly
"setShelter") and "shelteredValue". The symbol table used to
be sheltered in the continuation field of the root
threadData vector, and is now kept in the hidden entity
field of the internals object.