Code cleanup.
At some point the prepareMessageReturn() macro was changed
so that, upon discovering that it was keeping a promise,
it would explicitly end the thread rather than allowing the
function return to do this (probably due to a change in
where tailcalls were being expected). At that point it no
longer actually needed to be a macro, and now it isn't.