Better seperation between interactive and batch mode.
When Gospel is run with a command-line argument, it now
executes that file and then exits, with a default exception
handler that also exits. When run without a command-line
argument it runs "REPL.gs", which presents a REPL and sets
up a default exception handler to return to the REPL. Syntax
error line-number reporting has been disabled until it can
be reimplemented in accord with the new system.