Fixed resource leak in call().
Previously, if call() caused an exception to be raised, the
interpreter would continue in the exception's thread, with
the thread that invoked call() forever waiting for its
promise to be fulfilled. Now, call() installs an exception
handler that fulfills the promise with a special object that
call() can test for, to indicate that call() should
terminate the thread that invoked it.