Renaming object#new to object#instance.
It is intended that the normal convention be for prototypes
to redefine $new in whatever was is appropriate to create a
new instance of the type they represent. There needs to be
another method that will always mean "create a new object
with the target as its proto". This message is $instance.