Reimplementing the call() mechanism.
Instead of spawning a new thread (just to get a new C stack)
we now call doNext() again in the same thread, and
object#raise uses longjmp() to trim the stack down to the
context for the original invocation of doNext() when it's
time to deal with an exception. main() now creates an actor
to perform the initial loading of files, which means that
the parser now always runs in a complete VM environment and
doesn't need a special version of call(). Continuation
objects have a new field, used to shelter a previous
continuation from the garbage collector when native code is
calling back into the interpreter. Support for making
optimizable sibling calls to primitive methods has been
dropped: It was quite possibly a net performance loss, and
in any case was premature optimizaton that was unnecessarily
complicating the interpreter.