Decoupling of thread creation and actor creation.
Instead of being associated with a single thread for its
entire lifetime, waiting on a condition variable when its
queue is empty, an actor now kills its thread when it has
emptied its queue and spawns a new one next time it receives
a message. This eliminates the problem of how and when to
kill the thread of an actor that seems to be dead.