Further establishing the "atom" type.
For the time being, "atom" will be the unit for the length
of vectors. If we're serious about separating vectors of
atoms from vectors of references to save memory, the
allocator will have to be changed to deliver each type with
the correct alignment: Until then, we might as well say
"typedef intptr_t atom".